pyproject.toml
[project]
name = "pymcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.2.5",
    "mysql-connector-python>=9.3.0",
]
